+++ crates/node/ab-client-database/src/lib.rs
@@ -47,6 +47,7 @@
     default_field_values,
     get_mut_unchecked,
     iter_collect_into,
+    maybe_uninit_as_bytes,
     maybe_uninit_fill
 )]